pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "nodox"
version = "0.2.0"
description = "NoDox - Neutralizing Doxware & Modern Ransomware Extortion"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "NoDox Team", email = "nodox@example.com"}
]
keywords = [
    "security",
    "ransomware",
    "doxware",
    "encryption",
    "data-protection",
    "cybersecurity"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Information Technology",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
    "Topic :: Security :: Cryptography",
    "Topic :: System :: Systems Administration",
]
dependencies = [
    "cryptography>=43.0.0",
    "pyyaml>=6.0.1",
    "psutil>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/Orantes67/NoDox"
Documentation = "https://github.com/Orantes67/NoDox#readme"
Repository = "https://github.com/Orantes67/NoDox.git"
Issues = "https://github.com/Orantes67/NoDox/issues"

[project.scripts]
nodox = "nodox.nodox:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["nodox*"]

[tool.setuptools.package-data]
nodox = ["config/*.yaml", "config/*.txt"]

[tool.black]
line-length = 100
target-version = ['py310', 'py311', 'py312']

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=nodox --cov-report=html"
